WIP: Trajectory Publishing (#52)
* Create soccer_trajectories package

* Add script to publish dummy trajectory to robot

* Add folder for trajectories

* Move trajectories_publisher into its own module, shorten script

* Rename scripts as nodes

* Fix package configuration

* Remove extra files

* Add publishing from CSV and example trajectory

* Move nodes to scripts

* Update trajectories

* add dependacy on scipy and remove autogenerated comments

* add namespace so that commands are received by robot1

* Fixed navigation and publish trajectories

* Fixed navigation and publish trajectories

* Updated motor mapping

* undo change

* Add some bug fixes

* fix trajectories at the end so that robot stays in squat position

+++ soccer_hardware/src/soccer_hardware/communication.py
@@ -5,7 +5,7 @@ from prettytable import PrettyTable
 from transmitter import Transmitter
 from receiver import Receiver
 from utility import *
-from std_msgs.msg import Float32
+from std_msgs.msg import Float64
 from control_msgs.msg import JointControllerState
 from sensor_msgs.msg import Imu
 from geometry_msgs.msg import Vector3
@@ -35,7 +35,7 @@ class Communication:
         self._pub_imu = rp.Publisher('imu', Imu, queue_size=1)
         self._motor_map = rp.get_param("~motor_mapping")
         for motor in self._motor_map:
-            self._motor_map[motor]["subscriber"] = rp.Subscriber(motor + "/command", Float32, self.trajectory_callback, motor)
+            self._motor_map[motor]["subscriber"] = rp.Subscriber(motor + "/command", Float64, self.trajectory_callback, motor)
             self._motor_map[motor]["publisher"] = rp.Publisher(motor + "/state", JointControllerState, queue_size=1)
             self._motor_map[motor]["value"] = 0.0
